+++ Attendance/src/main/java/com/system/Attendance/repository/MembersRepository.java
@@ -3,6 +3,6 @@ package com.system.Attendance.repository;
 import com.system.Attendance.domain.Member;
 import edu.miu.common.repository.BaseRepository;
 
-public interface MembersRepository extends BaseRepository<Member, Integer> {
+public interface MembersRepository extends BaseRepository<Member, Long> {
 
 }
